one more docs fix
svn:sync-xref-src-repo-rev: r69726

+++ src/VBox/Frontends/VBoxManage/VBoxManage.h
@@ -1,4 +1,4 @@
-/* $Id: VBoxManage.h 35761 2011-01-28 13:19:26Z noreply@oracle.com $ */
+/* $Id: VBoxManage.h 35764 2011-01-28 13:56:15Z noreply@oracle.com $ */
 /** @file
  * VBoxManage - VirtualBox command-line interface, internal header file.
  */
@@ -162,7 +162,7 @@ int handleInternalCommands(HandlerArg *a);
 
 /* VBoxManageControlVM.cpp */
 int handleControlVM(HandlerArg *a);
-ULONG getMaxNics(IVirtualBox* vbox, IMachine* mach);
+unsigned int getMaxNics(IVirtualBox* vbox, IMachine* mach);
 
 /* VBoxManageModifyVM.cpp */
 int handleModifyVM(HandlerArg *a);
